% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsirm_mle_functions.R
\name{data_generation}
\alias{data_generation}
\title{Data Generation}
\usage{
data_generation(seed = 1, N, nitem = NULL, gamma = 0.5, item = NULL)
}
\arguments{
\item{seed}{seed number for replication}

\item{N}{the number of respondents}

\item{nitem}{the number of items}

\item{gamma}{gamma parameter of LSIRM}
}
\description{
Data Generation
}
