% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsirm_mle_functions.R
\name{lsirm}
\alias{lsirm}
\title{ML estimation of LSIRM}
\usage{
lsirm(
  data,
  dimension = 3,
  model = 1,
  range = c(-4, 4),
  q = 11,
  max_iter = 500,
  threshold = 0.001
)
}
\arguments{
\item{data}{data matrix}

\item{dimension}{the number of dimension (main + interaction). The default is 3.}

\item{range}{range of the quadrature points}

\item{q}{the number of quadrature points}

\item{max_iter}{the maximum number of EM iteration}

\item{threshold}{the threshold to determine the EM convergence}
}
\description{
ML estimation of LSIRM
}
\examples{
\dontrun{
################################################################################
# SIMULATION DATA
################################################################################

# data generation
dataset <- data_generation(seed = 1234,
                           N = 2000,
                           nitem = 10,
                           gamma = 1.0)

data <- dataset$data
item <- dataset$item
theta <- dataset$theta

# model fitting
fit <- lsirm(data,
             max_iter = 200,
             threshold = 0.001)

# plotting
plot.lsirm(item)
plot.lsirm(fit$par_est)
plot.lsirm(fit)

plot(item[,1], fit$par_est[,1])
abline(0,1)

################################################################################
# DRV DATA
################################################################################
drv_data <- read.table("drv_data.txt")
fit.drv <- lsirm(drv_data,
                 max_iter = 200,
                 threshold = 0.001)

plot.lsirm(fit.drv)
}
}
