% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsirm_mle_functions.R
\name{plot.lsirm}
\alias{plot.lsirm}
\title{Title}
\usage{
\method{plot}{lsirm}(item, range = c(-3, 3))
}
\arguments{
\item{item}{item parameters, or a list object from \code{lsirm}}

\item{range}{range of the coordinates}
}
\description{
Title
}
